+++ gazebo/test/ServerFixture.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2012-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -509,8 +509,8 @@ void ServerFixture::GetFrame(const std::string &_cameraName,
   sensors::CameraSensorPtr camSensor =
     std::dynamic_pointer_cast<sensors::CameraSensor>(sensor);
 
-  _width = camSensor->GetImageWidth();
-  _height = camSensor->GetImageHeight();
+  _width = camSensor->ImageWidth();
+  _height = camSensor->ImageHeight();
 
   if (*_imgData)
   {
@@ -522,14 +522,14 @@ void ServerFixture::GetFrame(const std::string &_cameraName,
 
   this->gotImage = 0;
   event::ConnectionPtr c =
-    camSensor->GetCamera()->ConnectNewImageFrame(
+    camSensor->Camera()->ConnectNewImageFrame(
         boost::bind(&ServerFixture::OnNewFrame,
                     this, _1, _2, _3, _4, _5));
 
   while (this->gotImage < 20)
     common::Time::MSleep(100);
 
-  camSensor->GetCamera()->DisconnectNewImageFrame(c);
+  camSensor->Camera()->DisconnectNewImageFrame(c);
 }
 
 /////////////////////////////////////////////////